update PDF processing script to handle PDFs from a directory and to convert multiple JSONs to CSVs
