+++ .vscode/launch.json
@@ -1,24 +1,46 @@
 {
-    "configurations": [
-      {
-        "type": "debugpy",
-        "request": "launch",
-        "name": "Launch Program",
-        "program": "${workspaceFolder}/${input:programName}",
-        "args": ["${input:pdfFileName}"]
-      }
-    ],
-    "inputs": [
-      {
-        "type": "pickString",
-        "id": "programName",
-        "description": "Select the Python file to debug",
-        "options": ["chunking.py", "json_to_csv.py", "synth-openai.py"]
-      },
-      {
-        "type": "promptString",
-        "id": "pdfFileName",
-        "description": "Enter the PDF file name"
-      }
-    ]
-  }
+  "configurations": [
+    {
+      "name": "Python Debugger: Current File with Arguments",
+      "type": "debugpy",
+      "request": "launch",
+      "program": "${file}",
+      "console": "integratedTerminal",
+      "args": [
+        "${command:pickArgs}"
+      ]
+    },
+    {
+      "type": "debugpy",
+      "request": "launch",
+      "name": "Launch Program",
+      "program": "${workspaceFolder}/${input:programName}",
+      "args": [
+        "${input:pdfFileName}"
+      ]
+    }
+  ],
+  "inputs": [
+    {
+      "type": "pickString",
+      "id": "programName",
+      "description": "Select the Python file to debug",
+      "options": [
+        "chunking.py",
+        "generate-dataset-from-pdfs.py"
+      ],
+      "default": "generate-dataset-from-pdfs.py"
+
+    },
+    {
+      "type": "promptString",
+      "id": "pdfFileName",
+      "description": "Enter the PDF file name"
+    },
+    {
+      "type": "promptString",
+      "id": "pdfDirectory",
+      "description": "Enter the PDF directory"
+    }
+  ]
+}
