+++ output_datasets/A_brief_history_of_art.csv
@@ -1,0 +1,6 @@
+prompt,question,answer
+You are a helpful assistant who can answer questions about the topic in the dataset.,What is the primary characteristic of classical Greek Art?,Heroic realism
+You are a helpful assistant who can answer questions about the topic in the dataset.,What architectural styles and innovations emerged during the Roman period (850 BC to 476 AD)?,"The Romans adapted architecture to their pragmatic way of life, developing new complex constructions with achievements in engineering like the arch, vault, and dome, as well as the invention of concrete facilitating public structures like aqueducts, baths, and amphitheatres."
+You are a helpful assistant who can answer questions about the topic in the dataset.,What were the characteristics of the Baroque style in art?,"The Baroque style was characterized by sculptures with richly decorated details, impressive statues and structures with grandiose movements and postures like a theatrical performance. It was often represented as a second phase of Renaissance, with an emphasis on grandeur and drama."
+You are a helpful assistant who can answer questions about the topic in the dataset.,What was Art Nouveau?,"Art Nouveau was an international trend in the decorative and fine arts that developed between about 1880 and 1920, focusing on traditional craftsmanship and often using medieval, romantic, or folk styles of decoration."
+You are a helpful assistant who can answer questions about the topic in the dataset.,What is Pop Art?,Pop Art (1950-1970) was a major post-war movement which exhibited mass-production merchandise with irony and fun.
