+++ output_datasets/Artificial Intelligence-Machine Learning Explained.json
@@ -1,0 +1,48 @@
+{
+  "responses": [
+    {
+      "question": "What is one key difference between machine learning algorithms and classic computers?",
+      "answer": "Machine learning algorithms can adjust themselves without being explicitly programmed, unlike classic computers that require explicit programming."
+    },
+    {
+      "question": "What are the key differences between traditional programming and machine learning?",
+      "answer": "Traditional programming involves explicit rules, whereas machine learning teaches a computer to 'learn by example' through massive volumes of training data."
+    },
+    {
+      "question": "What are the main differences between training and inference in machine learning?",
+      "answer": "Inference requires substantially less compute power than training, but still benefits from specialized AI chips."
+    },
+    {
+      "question": "How will AI transform the threat landscape in terms of adversaries' capabilities?",
+      "answer": "AI will enable adversaries to act with micro-precision at a macro-scale and with greater speed, allowing them to target individuals in new ways and manipulate AI systems that citizens rely upon."
+    },
+    {
+      "question": "Has open-source code libraries and developer tools made machine learning more accessible to organizations?",
+      "answer": "yes"
+    },
+    {
+      "question": "What are some limitations of current AI technology?",
+      "answer": "Current AI technology struggles with tasks such as generalizing well to new data, estimating uncertainty/confidence, choosing its own goals, figuring out cause and effect or why something happened, thinking creatively or applying common sense, creating a strategy, and lacks generalized intelligence."
+    },
+    {
+      "question": "What is the primary obstacle to innovation in national security?",
+      "answer": "The culture within institutions, specifically the DoD and IC, being resistant to change and reliant on traditional methods."
+    },
+    {
+      "question": "What are the main differences between supervised learning, unsupervised learning, semi-supervised learning, and reinforcement learning?",
+      "answer": "Supervised learning algorithms fall into four classes: Supervised Learning, Unsupervised Learning, Semi-supervised Learning, and Reinforcement Learning. They differ based on what types of data their algorithms can work with, whether or not the training data is labeled, and how the system receives its data inputs."
+    },
+    {
+      "question": "What are the four broad families of Machine Learning algorithms?",
+      "answer": "The four major families of algorithms are Supervised Learning, Unsupervised Learning, Semi-Supervised Learning, and Reinforcement Learning."
+    },
+    {
+      "question": "What is a key use case for unsupervised learning?",
+      "answer": "Fraud detection in financial data"
+    },
+    {
+      "question": "What is Semi-Supervised Learning?",
+      "answer": "A class of algorithms that combine techniques from Supervised and Unsupervised learning for applications with a small set of labeled data and a large set of unlabeled data."
+    }
+  ]
+}
