--- responses.csv
@@ -1,14 +1,0 @@
-prompt,question,answer
-You will answer questions about Machine Learning,What is the significance of Artificial Intelligence and Machine Learning in the context of national security and innovation?,"Artificial Intelligence/Machine Learning – Explained: Hundreds of billions in public and private capital is being invested in AI and Machine Learning companies, which are transforming businesses and government agencies with technologies that will change the balance of military power."
-You will answer questions about Machine Learning,What is machine learning?,Machine learning algorithms - computer programs that adjust themselves to perform better as they are exposed to more data.
-You will answer questions about Machine Learning,What are the main differences between traditional programming and machine learning?,"Traditional programming involves writing explicit rules, whereas machine learning trains a computer to learn by example through massive volumes of training data."
-You will answer questions about Machine Learning,Why do machine learning models need to be updated regularly?,Machine learning models get stale over time if they are not updated regularly with new training data that matches the changing state of the world.
-You will answer questions about Machine Learning,How are applications using artificial intelligence and machine learning?,"Applications are already appearing in virtual customer assistants with speech recognition, sentiment analysis, automated/augmented quality assurance and other technologies providing customers with 24/7 self- and assisted-service options across channels."
-You will answer questions about Machine Learning,What are the main factors contributing to the increasing threat landscape due to AI?,"Four key changes make machine learning possible now: 1) massive data sets, 2) improved machine learning algorithms, 3) open-source code, pretrained models and frameworks, and 4) more computing power."
-You will answer questions about Machine Learning,What are some limitations of AI?,"AI can be easily fooled by out-of-domain data, overfitting, underfitting, and lacks generalized intelligence, creativity, and common sense."
-You will answer questions about Machine Learning,What is the primary obstacle to innovation in national security?,"The primary obstacle to innovation in national security is not technology, but culture."
-You will answer questions about Machine Learning,What are some examples of companies designing their own AI chips?,"Companies like Facebook, Google, Amazon, BrainChip, GrAI Matter, Innatera, Intel, Mythic, Aspinity, Tetramem, Lightmatter, Lightelligence, Luminous, Lighton, Nvidia, AMD, Qualcomm, Hailo-8, Ambarella CV5S, Quadric (Q16), RealTek 3916N, Syntiant (NDP)"
-You will answer questions about Machine Learning,What is the main challenge faced by new AI Accelerator vendors in porting popular AI applications to their chips?,"They have to build their own drivers, compiler, debugger, and other tools due to Nvidia's software moat around their GPU's."
-You will answer questions about Machine Learning,What are the different types of Machine Learning?,"There are four main categories: Supervised Learning, Unsupervised Learning, Semi-Supervised Learning, and Reinforcement Learning."
-You will answer questions about Machine Learning,What is unsupervised learning?,Unsupervised learning is a method of machine learning where models are trained on large amounts of unclassified data to identify patterns and relationships without human intervention.
-You will answer questions about Machine Learning,What is the key difference between Supervised and Reinforcement Learning in Artificial Intelligence?,"Reinforcement Learning gathers data by an autonomous AI agent, whereas Supervised Learning relies on pre-labeled data."
