--- responses.json
@@ -1,56 +1,0 @@
-{
-  "responses": [
-    {
-      "question": "What is the significance of Artificial Intelligence and Machine Learning in the context of national security and innovation?",
-      "answer": "Artificial Intelligence/Machine Learning \u2013 Explained: Hundreds of billions in public and private capital is being invested in AI and Machine Learning companies, which are transforming businesses and government agencies with technologies that will change the balance of military power."
-    },
-    {
-      "question": "What is machine learning?",
-      "answer": "Machine learning algorithms - computer programs that adjust themselves to perform better as they are exposed to more data."
-    },
-    {
-      "question": "What are the main differences between traditional programming and machine learning?",
-      "answer": "Traditional programming involves writing explicit rules, whereas machine learning trains a computer to learn by example through massive volumes of training data."
-    },
-    {
-      "question": "Why do machine learning models need to be updated regularly?",
-      "answer": "Machine learning models get stale over time if they are not updated regularly with new training data that matches the changing state of the world."
-    },
-    {
-      "question": "How are applications using artificial intelligence and machine learning?",
-      "answer": "Applications are already appearing in virtual customer assistants with speech recognition, sentiment analysis, automated/augmented quality assurance and other technologies providing customers with 24/7 self- and assisted-service options across channels."
-    },
-    {
-      "question": "What are the main factors contributing to the increasing threat landscape due to AI?",
-      "answer": "Four key changes make machine learning possible now: 1) massive data sets, 2) improved machine learning algorithms, 3) open-source code, pretrained models and frameworks, and 4) more computing power."
-    },
-    {
-      "question": "What are some limitations of AI?",
-      "answer": "AI can be easily fooled by out-of-domain data, overfitting, underfitting, and lacks generalized intelligence, creativity, and common sense."
-    },
-    {
-      "question": "What is the primary obstacle to innovation in national security?",
-      "answer": "The primary obstacle to innovation in national security is not technology, but culture."
-    },
-    {
-      "question": "What are some examples of companies designing their own AI chips?",
-      "answer": "Companies like Facebook, Google, Amazon, BrainChip, GrAI Matter, Innatera, Intel, Mythic, Aspinity, Tetramem, Lightmatter, Lightelligence, Luminous, Lighton, Nvidia, AMD, Qualcomm, Hailo-8, Ambarella CV5S, Quadric (Q16), RealTek 3916N, Syntiant (NDP)"
-    },
-    {
-      "question": "What is the main challenge faced by new AI Accelerator vendors in porting popular AI applications to their chips?",
-      "answer": "They have to build their own drivers, compiler, debugger, and other tools due to Nvidia's software moat around their GPU's."
-    },
-    {
-      "question": "What are the different types of Machine Learning?",
-      "answer": "There are four main categories: Supervised Learning, Unsupervised Learning, Semi-Supervised Learning, and Reinforcement Learning."
-    },
-    {
-      "question": "What is unsupervised learning?",
-      "answer": "Unsupervised learning is a method of machine learning where models are trained on large amounts of unclassified data to identify patterns and relationships without human intervention."
-    },
-    {
-      "question": "What is the key difference between Supervised and Reinforcement Learning in Artificial Intelligence?",
-      "answer": "Reinforcement Learning gathers data by an autonomous AI agent, whereas Supervised Learning relies on pre-labeled data."
-    }
-  ]
-}
